+++ dialogflow_v2/types.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright 2019 Google LLC
+# Copyright 2020 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -27,9 +27,11 @@ from dialogflow_v2.proto import entity_type_pb2
 from dialogflow_v2.proto import intent_pb2
 from dialogflow_v2.proto import session_entity_type_pb2
 from dialogflow_v2.proto import session_pb2
+from dialogflow_v2.proto import validation_result_pb2
 from dialogflow_v2.proto import webhook_pb2
 from google.longrunning import operations_pb2
 from google.protobuf import any_pb2
+from google.protobuf import duration_pb2
 from google.protobuf import empty_pb2
 from google.protobuf import field_mask_pb2
 from google.protobuf import struct_pb2
@@ -40,6 +42,7 @@ from google.type import latlng_pb2
 _shared_modules = [
     operations_pb2,
     any_pb2,
+    duration_pb2,
     empty_pb2,
     field_mask_pb2,
     struct_pb2,
@@ -55,6 +58,7 @@ _local_modules = [
     intent_pb2,
     session_entity_type_pb2,
     session_pb2,
+    validation_result_pb2,
     webhook_pb2,
 ]